# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: QueueMate API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /users/me:
    get:
      summary: Current user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
    patch:
      summary: Update current user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '409': { description: Nickname already taken }
  /users/me/game-accounts:
    get:
      summary: My linked game accounts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/GameAccountView' }
    post:
      summary: Link a game account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGameAccountRequest' }
      responses:
        '201':
          description: Linked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameAccountView' }
        '409': { description: Already linked }
  /users/me/game-accounts/{id}:
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      summary: Unlink a game account
      responses:
        '204': { description: Unlinked }
        '404': { description: Not found }
...
  /blocks:
    get:
      summary: Block list
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BlockView' }
    post:
      summary: Block user
      description: |
        차단하면 기존 친구 관계와 대기 중인 친구 요청이 함께 제거되고,
        이후 매칭 후보에서 양방향으로 즉시 제외된다 (docs/01 Block flow, INV-6).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBlockRequest' }
      responses:
        '201':
          description: Blocked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlockView' }
        '409': { description: Already blocked or self }
  /blocks/{userId}:
    delete:
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      summary: Unblock user
      responses:
        '204': { description: Unblocked }
        '404': { description: Not blocked }
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, nickname]
      properties:
        id: { type: string, format: uuid }
        nickname: { type: string }
        avatarUrl: { type: string, nullable: true }
    UpdateUserRequest:
      type: object
      properties:
        nickname: { type: string, minLength: 2, maxLength: 16 }
        avatarUrl: { type: string, nullable: true }
    GameAccountView:
      type: object
      required: [id, game, externalGameId]
      properties:
        id: { type: string, format: uuid }
        game: { $ref: '#/components/schemas/GameKey' }
        externalGameId: { type: string }
        region: { type: string, nullable: true }
        rankCode: { type: string, nullable: true }
        verifiedAt: { type: string, format: date-time, nullable: true }
    CreateGameAccountRequest:
      type: object
      required: [game, externalGameId]
      properties:
        game: { $ref: '#/components/schemas/GameKey' }
        externalGameId: { type: string, maxLength: 128 }
        region: { type: string, maxLength: 20, nullable: true }
    GameKey:
      type: string
      enum: [LOL, VALORANT, PUBG]
...
    BlockView:
      type: object
      required: [userId]
      properties:
        userId: { type: string, format: uuid }
        nickname: { type: string }
        blockedAt: { type: string, format: date-time }
    CreateBlockRequest:
      type: object
      required: [targetUserId]
      properties:
        targetUserId: { type: string, format: uuid }
...